fix final bugs

+++ Server/Server/Item.h
@@ -10,13 +10,13 @@ struct Item
 	~Item() {};
 
 	void SetAliveState(bool alive) { itemInfo.set_isalive(alive); }
-	void SetBelongingId(uint32 id) { itemInfo.set_playerid(id); }
+	void SetBelongingId(uint64 id) { itemInfo.set_playerid(id); }
 	void SetPos(Vec2Int pos) { itemInfo.set_posx(pos.x); itemInfo.set_posy(pos.y); }
 	void SetItemType(Protocol::ITEM_TYPE type) { itemInfo.set_itemtype(type); }
-	uint32 GetItemID() { return itemInfo.itemid(); }
+	uint64 GetItemID() { return itemInfo.itemid(); }
 	Vec2Int GetCellPos() { return  Vec2Int(itemInfo.posx(), itemInfo.posy()); }
 	Protocol::ITEM_TYPE GetItemType() { return itemInfo.itemtype(); }
-	uint32 GetBelongingID() { return itemInfo.playerid(); }
+	uint64 GetBelongingID() { return itemInfo.playerid(); }
 
 	static Protocol::ITEM_TYPE GetRandomItemType();
 	//bool _isAlive=false;
